Resource Module added
