+++ src/main/java/com/pm/Project_Management_Server/Services/ResourceService.java
@@ -1,83 +1,20 @@
 package com.pm.Project_Management_Server.Services;
 
 import com.pm.Project_Management_Server.dto.ResourceDTO;
-import com.pm.Project_Management_Server.entity.Project;
-import com.pm.Project_Management_Server.entity.Resource;
-import com.pm.Project_Management_Server.entity.ResourceLevel;
-import com.pm.Project_Management_Server.Repositories.ProjectRepository;
-import com.pm.Project_Management_Server.Repositories.ResourceRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.BeanUtils;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-@Service
-@RequiredArgsConstructor
-public class ResourceService {
+public interface ResourceService {
 
-    private final ResourceRepository resourceRepository;
-    private final ProjectRepository projectRepository;
+    ResourceDTO addResource(ResourceDTO dto);
 
-    public ResourceDTO addResource(ResourceDTO dto) {
-        Project project = projectRepository.findById(dto.getProjectId())
-                .orElseThrow(() -> new RuntimeException("Project not found"));
+    List<ResourceDTO> getResourcesByProject(Long projectId);
 
-        Resource resource = new Resource();
-        BeanUtils.copyProperties(dto, resource);
-        resource.setProject(project);
-        resource.setAllocated(true);
+    List<ResourceDTO> getResourcesByLevel(String level);
 
-        Resource saved = resourceRepository.save(resource);
+    List<ResourceDTO> getAllocatedResources();
 
-        ResourceDTO response = new ResourceDTO();
-        BeanUtils.copyProperties(saved, response);
-        response.setProjectId(project.getId());
-        return response;
-    }
+    List<ResourceDTO> getUnallocatedResources();
 
-    public List<ResourceDTO> getResourcesByProject(Long projectId) {
-        return resourceRepository.findByProjectId(projectId)
-                .stream()
-                .map(this::convertToDTO)
-                .collect(Collectors.toList());
-    }
-
-    public List<ResourceDTO> getResourcesByLevel(String level) {
-        ResourceLevel resourceLevel = ResourceLevel.valueOf(level.toUpperCase());
-        return resourceRepository.findByLevel(resourceLevel)
-                .stream()
-                .map(this::convertToDTO)
-                .collect(Collectors.toList());
-    }
-
-    public List<ResourceDTO> getAllocatedResources() {
-        return resourceRepository.findByAllocatedTrue()
-                .stream()
-                .map(this::convertToDTO)
-                .collect(Collectors.toList());
-    }
-
-    public List<ResourceDTO> getUnallocatedResources() {
-        return resourceRepository.findByAllocatedFalse()
-                .stream()
-                .map(this::convertToDTO)
-                .collect(Collectors.toList());
-    }
-
-    public ResourceDTO updateResource(Long id, ResourceDTO dto) {
-        Resource resource = resourceRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Resource not found"));
-        BeanUtils.copyProperties(dto, resource, "id", "project");
-        Resource updated = resourceRepository.save(resource);
-        return convertToDTO(updated);
-    }
-
-    private ResourceDTO convertToDTO(Resource resource) {
-        ResourceDTO dto = new ResourceDTO();
-        BeanUtils.copyProperties(resource, dto);
-        dto.setProjectId(resource.getProject().getId());
-        return dto;
-    }
+    ResourceDTO updateResource(Long id, ResourceDTO dto);
 }
