+++ src/main/java/com/pm/Project_Management_Server/controller/ResourceController.java
@@ -1,7 +1,8 @@
 package com.pm.Project_Management_Server.controller;
 
+
+import com.pm.Project_Management_Server.Services.ResourceServiceImpl;
 import com.pm.Project_Management_Server.dto.ResourceDTO;
-import com.pm.Project_Management_Server.Services.ResourceService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -13,7 +14,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ResourceController {
 
-    private final ResourceService resourceService;
+    private final ResourceServiceImpl resourceService;
 
     @PostMapping
     public ResponseEntity<ResourceDTO> createResource(@RequestBody ResourceDTO dto) {
