+++ src/main/java/com/pm/Project_Management_Server/services/OpenPositionServiceImpl.java
@@ -1,4 +1,4 @@
-package com.pm.Project_Management_Server.services;
+package com.pm.Project_Management_Server.Services;
 
 import com.pm.Project_Management_Server.dto.OpenPositionDTO;
 import com.pm.Project_Management_Server.entity.OpenPosition;
